Experiment acceration (#16)
* test: baseline batch

* docs: add experiment result

* test: tensorrt

* feat: onnx input name checker

* fix: trtexec

* docs: add experiment results

* fix: makefile

* fix: wrong baseline result bcz of torch sync

* feat: add result table and fix xformer benchamark code

* fix: xformer benchmark

* docs: xformer results

* docs: update experiment result

* docs: tensorrt plugin

* docs: add faster-trasnformer trt plugin result

* docs: result table update

* fix: typo

* test: xformers

* docs: result update

* docs: result update

* docs: update fp16 result

* feat: fp16 trt model

* docs: update fp16 result

* feat: fp16 baseline

* docs: update fp16 result

* docs: result update

* docs: add fp16 result

* docs: add dir table

+++ acceleration/00-vit-baseline/benchmark_fp16.py
@@ -15,11 +15,13 @@ if device == "cpu":
 # Define pretrained vit model
 model = timm.create_model("vit_base_patch16_224", pretrained=True)
 model = model.to(device)
+model.half()  # Convert to half precision
 
 # Define input
 batch_sizes = [1, 2, 4, 8, 16, 32, 64]
 for batch in batch_sizes:
     input = torch.randn(batch, 3, 224, 224)
+    input = input.half()  # Convert to half precision
 
     # Warm up
     for _ in range(10):
@@ -36,4 +38,6 @@ for batch in batch_sizes:
             end = time.time()
             inference_times.append((end - start) * 1000)
 
-    print(f"batch size :{batch}\nViT average inference time : {sum(inference_times)/len(inference_times)}ms")
+    print(
+        f"batch size :{batch}\nViT average inference time : {sum(inference_times)/len(inference_times)}ms"
+    )
